Add approved, approves to PRs
