+++ src/provider/__recordings__/fetchMergeRequestApprovals_2789555690/recording.har
@@ -1,0 +1,202 @@
+{
+  "log": {
+    "_recordingName": "fetchMergeRequestApprovals",
+    "creator": {
+      "comment": "persister:JupiterOneIntegationFSPersister",
+      "name": "Polly.JS",
+      "version": "4.1.0"
+    },
+    "entries": [
+      {
+        "_id": "12fb95c4be7c3b555992004b6ec8be97",
+        "_order": 0,
+        "cache": {},
+        "request": {
+          "bodySize": 0,
+          "cookies": [],
+          "headers": [
+            {
+              "_fromType": "array",
+              "name": "private-token",
+              "value": "[REDACTED]"
+            },
+            {
+              "_fromType": "array",
+              "name": "accept",
+              "value": "*/*"
+            },
+            {
+              "_fromType": "array",
+              "name": "user-agent",
+              "value": "node-fetch/1.0 (+https://github.com/bitinn/node-fetch)"
+            },
+            {
+              "_fromType": "array",
+              "name": "accept-encoding",
+              "value": "gzip,deflate"
+            },
+            {
+              "_fromType": "array",
+              "name": "connection",
+              "value": "close"
+            },
+            {
+              "name": "host",
+              "value": "gitlab.com"
+            }
+          ],
+          "headersSize": 273,
+          "httpVersion": "HTTP/1.1",
+          "method": "GET",
+          "queryString": [],
+          "url": "https://gitlab.com/api/v4/projects/19688677/merge_requests/1/approvals"
+        },
+        "response": {
+          "bodySize": 1031,
+          "content": {
+            "mimeType": "application/json",
+            "size": 1031,
+            "text": "{\"id\":78009276,\"iid\":1,\"project_id\":19688677,\"title\":\"WIP: Testing provided by Adam.\",\"description\":\"Reviewers should be aware that I'm just testing merge requests here.\",\"state\":\"opened\",\"created_at\":\"2020-11-13T22:45:44.886Z\",\"updated_at\":\"2020-11-13T22:54:55.261Z\",\"merge_status\":\"can_be_merged\",\"approved\":true,\"approvals_required\":0,\"approvals_left\":0,\"require_password_to_approve\":null,\"approved_by\":[{\"user\":{\"id\":7620849,\"name\":\"Adam Williams\",\"username\":\"adam_williams_trial_test\",\"state\":\"active\",\"avatar_url\":\"https://secure.gravatar.com/avatar/7db770e61e9d9eeefa1f9d89966dadd3?s=80\\u0026d=identicon\",\"web_url\":\"https://gitlab.com/adam_williams_trial_test\"}}],\"suggested_approvers\":[],\"approvers\":[],\"approver_groups\":[],\"user_has_approved\":false,\"user_can_approve\":false,\"approval_rules_left\":[],\"has_approval_rules\":false,\"merge_request_approvers_available\":true,\"multiple_approval_rules_available\":true}"
+          },
+          "cookies": [
+            {
+              "domain": ".gitlab.com",
+              "expires": "2020-12-14T02:38:29.000Z",
+              "httpOnly": true,
+              "name": "__cfduid",
+              "path": "/",
+              "sameSite": "Lax",
+              "secure": true,
+              "value": "[REDACTED]"
+            }
+          ],
+          "headers": [
+            {
+              "name": "date",
+              "value": "Sat, 14 Nov 2020 02:38:29 GMT"
+            },
+            {
+              "name": "content-type",
+              "value": "application/json"
+            },
+            {
+              "name": "connection",
+              "value": "close"
+            },
+            {
+              "_fromType": "array",
+              "name": "set-cookie",
+              "value": "[REDACTED]"
+            },
+            {
+              "name": "vary",
+              "value": "Accept-Encoding, Origin"
+            },
+            {
+              "name": "cache-control",
+              "value": "max-age=0, private, must-revalidate"
+            },
+            {
+              "name": "etag",
+              "value": "W/\"e62406e4deb543f9116f8997bde366a8\""
+            },
+            {
+              "name": "x-content-type-options",
+              "value": "nosniff"
+            },
+            {
+              "name": "x-frame-options",
+              "value": "SAMEORIGIN"
+            },
+            {
+              "name": "x-gitlab-feature-category",
+              "value": "code_review"
+            },
+            {
+              "name": "x-request-id",
+              "value": "qeuw61qDsK6"
+            },
+            {
+              "name": "x-runtime",
+              "value": "0.066132"
+            },
+            {
+              "name": "strict-transport-security",
+              "value": "max-age=31536000"
+            },
+            {
+              "name": "referrer-policy",
+              "value": "strict-origin-when-cross-origin"
+            },
+            {
+              "name": "ratelimit-limit",
+              "value": "600"
+            },
+            {
+              "name": "ratelimit-observed",
+              "value": "4"
+            },
+            {
+              "name": "ratelimit-remaining",
+              "value": "596"
+            },
+            {
+              "name": "ratelimit-reset",
+              "value": "1605321569"
+            },
+            {
+              "name": "ratelimit-resettime",
+              "value": "Sat, 14 Nov 2020 02:39:29 GMT"
+            },
+            {
+              "name": "gitlab-lb",
+              "value": "fe-08-lb-gprd"
+            },
+            {
+              "name": "gitlab-sv",
+              "value": "localhost"
+            },
+            {
+              "name": "cf-cache-status",
+              "value": "DYNAMIC"
+            },
+            {
+              "name": "cf-request-id",
+              "value": "0666376389000074651dac9000000001"
+            },
+            {
+              "name": "expect-ct",
+              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
+            },
+            {
+              "name": "server",
+              "value": "cloudflare"
+            },
+            {
+              "name": "cf-ray",
+              "value": "5f1d5b4c0f0a7465-IAD"
+            }
+          ],
+          "headersSize": 1102,
+          "httpVersion": "HTTP/1.1",
+          "redirectURL": "",
+          "status": 200,
+          "statusText": "OK"
+        },
+        "startedDateTime": "2020-11-14T02:38:29.718Z",
+        "time": 178,
+        "timings": {
+          "blocked": -1,
+          "connect": -1,
+          "dns": -1,
+          "receive": 0,
+          "send": 0,
+          "ssl": -1,
+          "wait": 178
+        }
+      }
+    ],
+    "pages": [],
+    "version": "1.2"
+  }
+}
