+++ src/steps/build-user-approved-merge-request-relationships/index.ts
@@ -75,7 +75,7 @@ async function createUserIdMap(
   return userIdMap;
 }
 
-export default createStep((instance) => createGitlabClient(instance));
+export default createStep((instance) => createGitlabClient(instance.config));
 
 export function createUserApprovedPrRelationship(
   user: Entity,
