--- test/context.ts
@@ -1,1 +1,0 @@
-export { createMockStepExecutionContext } from '@jupiterone/integration-sdk-testing';
