+++ test/fixtures/mergeRequests.json
@@ -1,0 +1,116 @@
+[
+  {
+    "mergeRequest": {
+      "id": 78009276,
+      "iid": 1,
+      "project_id": 19688677,
+      "title": "WIP: Testing provided by Adam.",
+      "description": "Reviewers should be aware that I'm just testing merge requests here.",
+      "state": "opened",
+      "created_at": "2020-11-13T22:45:44.886Z",
+      "updated_at": "2020-11-13T22:54:55.261Z",
+      "merged_by": null,
+      "merged_at": null,
+      "closed_by": null,
+      "closed_at": null,
+      "target_branch": "master",
+      "source_branch": "testing-merge-requests",
+      "user_notes_count": 0,
+      "upvotes": 0,
+      "downvotes": 0,
+      "author": {
+        "id": 6290829,
+        "name": "Adam Williams",
+        "username": "adamiwilliams",
+        "state": "active",
+        "avatar_url": "https://secure.gravatar.com/avatar/a5877d701b5d6af7f990ced46067ca85?s=80\\u0026d=identicon",
+        "web_url": "https://gitlab.com/adamiwilliams"
+      },
+      "assignees": [
+        {
+          "id": 6290829,
+          "name": "Adam Williams",
+          "username": "adamiwilliams",
+          "state": "active",
+          "avatar_url": "https://secure.gravatar.com/avatar/a5877d701b5d6af7f990ced46067ca85?s=80\\u0026d=identicon",
+          "web_url": "https://gitlab.com/adamiwilliams"
+        }
+      ],
+      "assignee": {
+        "id": 6290829,
+        "name": "Adam Williams",
+        "username": "adamiwilliams",
+        "state": "active",
+        "avatar_url": "https://secure.gravatar.com/avatar/a5877d701b5d6af7f990ced46067ca85?s=80\\u0026d=identicon",
+        "web_url": "https://gitlab.com/adamiwilliams"
+      },
+      "source_project_id": 19688677,
+      "target_project_id": 19688677,
+      "labels": [],
+      "work_in_progress": true,
+      "milestone": null,
+      "merge_when_pipeline_succeeds": false,
+      "merge_status": "can_be_merged",
+      "sha": "ba40cb232aba589cd0fff308802542cf2b07f619",
+      "merge_commit_sha": null,
+      "squash_commit_sha": null,
+      "discussion_locked": null,
+      "should_remove_source_branch": null,
+      "force_remove_source_branch": true,
+      "reference": "!1",
+      "references": {
+        "short": "!1",
+        "relative": "!1",
+        "full": "team-jupiterone/j1-gitbook-docs!1"
+      },
+      "web_url": "https://gitlab.com/team-jupiterone/j1-gitbook-docs/-/merge_requests/1",
+      "time_stats": {
+        "time_estimate": 0,
+        "total_time_spent": 0,
+        "human_time_estimate": null,
+        "human_total_time_spent": null
+      },
+      "squash": false,
+      "task_completion_status": { "count": 0, "completed_count": 0 },
+      "has_conflicts": false,
+      "blocking_discussions_resolved": true,
+      "approvals_before_merge": null
+    },
+    "approval": {
+      "id": 78009276,
+      "iid": 1,
+      "project_id": 19688677,
+      "title": "WIP: Testing provided by Adam.",
+      "description": "Reviewers should be aware that I'm just testing merge requests here.",
+      "state": "opened",
+      "created_at": "2020-11-13T22:45:44.886Z",
+      "updated_at": "2020-11-13T22:54:55.261Z",
+      "merge_status": "can_be_merged",
+      "approved": true,
+      "approvals_required": 0,
+      "approvals_left": 0,
+      "require_password_to_approve": null,
+      "approved_by": [
+        {
+          "user": {
+            "id": 7620849,
+            "name": "Adam Williams",
+            "username": "adam_williams_trial_test",
+            "state": "active",
+            "avatar_url": "https://secure.gravatar.com/avatar/7db770e61e9d9eeefa1f9d89966dadd3?s=80\\u0026d=identicon",
+            "web_url": "https://gitlab.com/adam_williams_trial_test"
+          }
+        }
+      ],
+      "suggested_approvers": [],
+      "approvers": [],
+      "approver_groups": [],
+      "user_has_approved": false,
+      "user_can_approve": false,
+      "approval_rules_left": [],
+      "has_approval_rules": false,
+      "merge_request_approvers_available": true,
+      "multiple_approval_rules_available": true
+    }
+  }
+]
