+++ test/index.ts
@@ -1,2 +1,1 @@
-export * from './context';
 export * from './recording';
